requirements.txt
streamlit
paddleocr
paddlepaddle
<<<<<<< HEAD
opencv-python-headless
=======
>>>>>>> f4b8613 (🔧 Removed OpenCV to fix Streamlit Cloud deployment)
numpy
pillow
pandas
rapidfuzz
